feat: get customer by id #4

+++ src/main/java/pl/codehouse/nn/bank/customer/CustomerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface CustomerRepository extends ReactiveCrudRepository<Customer, Integer> {
+interface CustomerRepository extends ReactiveCrudRepository<Customer, Long> {
 
 }
